+++ example/example1.js
@@ -9,8 +9,39 @@ var pannelLog = require('../index') ;
     setTimeout(pannelLog.setPercentComplete, 2000, 0.8);
     setTimeout(pannelLog.setPercentComplete, 2500, 1);
     //Your custom data
+    function getBigData(){
+        var arr = [];
+        for(var i = 0; i < 10000; i++){
+            arr.push({x:Math.random(), y:Math.random()});
+        }
+        return arr;
+    }
+    var infiniteArray = [];
+    pannelLog.addLineGroup(2, "TESTS HERE") ;
+    //make itens with method
     
+    
+    let count = 0;
+    pannelLog.addItem(2, 2, (colorInfo)=>{
+        if(count%2 == 0){
+            colorInfo.color = [pannelLog.color.yellow];
+        }
+        return "TEST 1" ;
+    }, 20, 44);
+    pannelLog.addItem(2, 3, "TEST 2", 20, (colorInfo)=>{
+        colorInfo.color = [pannelLog.color.blue];
+        return count++;
+    })
+
+
+    //make lines with your own
     pannelLog.onUpdate.add(()=>{
+        if(pannelLog.memoryUsage < 1200){
+            //spending memory to test log
+            for(var i = 0; i < 100; i++){
+                infiniteArray.push({x:Math.random(), y:Math.random(), i, data:getBigData()});
+            }
+        }
         //my list
         var myList = [{x:1, value:3},{x:2, value:3},{x:3, value:3},{x:4, value:3},{x:5, value:3}] ;
         //create a line